# add StepRange support for CartesianIndices