This PR intended to only add StepRange support for CartesianIndices, but more things are added here as I was working on it.

Three major changes are involved here:

## StepRange CartesianIndices

The biggest change here is to make it possible to pass a `step` to CartesianIndices so that more flexible loop behavior can be done

```julia
R = CartesianIndices(A)
Δ = CartesianIndex(ntuple(_->2, ndims(A)))

for I in first(R):Δ:last(R)
    ...
end
```

Another new thing introduced by this PR is that `step(::CartesianIndices)` now returns a `CartesianIndex`, this enables us to treat CartesianIndices as a multidimensional range type.

## Reduce CartesianIndices overhead

By reworking the iterating part, the overhead issue #9080 is accidentally fixed.

```julia
using Test
using BenchmarkTools

function sumcart_iter(A)
    s = zero(eltype(A))
    for I in CartesianIndices(A)
        @inbounds s += A[I]
    end
    s
end
function sumcart_manual(A::AbstractMatrix)
    s = zero(eltype(A))
    ax1, ax2 = axes(A)
    for j in ax2, i in ax1
        @inbounds s += A[i, j]
    end
    s
end


```

```julia
julia> A = rand(4, 4);

julia> @btime sumcart_iter($A)
  6.757 ns (0 allocations: 0 bytes) # this PR
  15.315 ns (0 allocations: 0 bytes) # master

julia> @btime sumcart_manual($A) # baseline
  8.013 ns (0 allocations: 0 bytes)

julia> A = rand(256,20);

julia> @btime sumcart_iter($A);
  4.543 μs (0 allocations: 0 bytes) # this PR
  5.017 μs (0 allocations: 0 bytes) # master

julia> @btime sumcart_manual($A);
  4.544 μs (0 allocations: 0 bytes) # baseline
```

So :tada: looping over `CartesianIndices` is marginally faster than a manual loop.

There's still some overhead for StepRange:

```julia
function sumcart_iter(A)
    s = zero(eltype(A))
    R = CartesianIndices(A)
    Δ = CartesianIndex(ntuple(_->2, ndims(R)))
    for I in first(R):Δ:last(R)
        @inbounds s += A[I]
    end
    s
end

function sumcart_manual(A::AbstractMatrix)
    s = zero(eltype(A))
    ax1, ax2 = axes(A)
    for j in first(ax2):2:last(ax2), i in first(ax1):2:last(ax1)
        @inbounds s += A[i, j]
    end
    s
end
```

```julia
julia> A = rand(8, 8);

julia> @btime sumcart_iter($A)
  36.418 ns (0 allocations: 0 bytes)

julia> @btime sumcart_manual($A)
  32.492 ns (0 allocations: 0 bytes)
```

but that's more difficult to deal with because (IIUC) `step` can be either positive or negative. I'm not very familiar with this kind of low-level optimization...

## tests and bug fixes

A lot of tests are spread in other files, I tried to collect most of the test patterns into `test/cartesian.jl` so it is easier to first check if `include("test/cartesian.jl")` works before triggering the whole heavy `make test` workflow.

A bug is found and fixed (see also #37928) :

```julia
# master
julia> CartesianIndices((2, 1:2))
1×2 CartesianIndices{2, Tuple{UnitRange{Int64}, UnitRange{Int64}}}:
 CartesianIndex(2, 1)  CartesianIndex(2, 2)

# this PR
julia> CartesianIndices((2, 1:2))
2×2 CartesianIndices{2, Tuple{Base.OneTo{Int64}, UnitRange{Int64}}}:
 CartesianIndex(1, 1)  CartesianIndex(1, 2)
 CartesianIndex(2, 1)  CartesianIndex(2, 2)

# master/this PR
julia> CartesianIndices((2, 2))
2×2 CartesianIndices{2, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}}}:
 CartesianIndex(1, 1)  CartesianIndex(1, 2)
 CartesianIndex(2, 1)  CartesianIndex(2, 2)
```


New tests are added to test common operations and properties. Old tests are still kept so as to not break anything unexpectedly.

---

There is still something I want to tackle, but I'm a bit running out of my spare time to keep working on this, which I'll leave them as future work:

- StepRange support to LinearIndices (https://github.com/JuliaLang/julia/pull/37829#discussion_r499183657)
- tweak the StepRange CartesianIndices overhead

cc: @timholy @mbauman 

closes #37577 
closes #9080