@johnnychen94 2020-10-04
Does `step(R::CartesianIndices) = CartesianIndex(step.(R.indices))` be crazy to you? With this PR done, I feel it might be pretty useful to think `CartesianIndices` as a multidimensional range now.