@johnnychen94 2020-10-07
@timholy I think this PR is ready for a hopefully final review

> One thing we can probably do to ease the transition is define
> ```julia
> const UnitCartesianIndices{N,R<:NTuple{N,AbstractUnitRange{Int}}} = CartesianIndices{N,R}
> ```
> and encourage people to use it when they want to enforce a step size of 1.

I finally choose to not add this because I didn't find a use case in this PR. Most methods are dispatched at the range type level and not the CartesianIndices type level.

Other noteworthy changes since the last review:

* I first implemented a [quite complicated overflow_check funcion](https://github.com/JuliaLang/julia/pull/37829/commits/4b47f634241b22e01bd8bbd3084d272033cf56dc#diff-89e97f67f8058d32eac8d917db5bbdbfR424-R433) and then reverted it to keep consistent with current range behavior. (CRef #37876)
* `step(::CartesianIndices)` is added