@timholy 2020-10-09
Let's find out if generalizing this type causes package problems: @nanosoldier `runtests(ALL, vs = ":master")`