@nanosoldier 2020-10-09
[Your package evaluation job](https://github.com/JuliaLang/julia/pull/37829#issuecomment-706126360) has completed - possible new issues were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/4086959a886ed94fbfd32417489e909b652a9952/pkgeval/by_hash/3386b42_vs_3559b72/report.md). cc @maleadt