@timholy 2020-10-09
Interesting failure in https://github.com/JuliaCI/NanosoldierReports/blob/4086959a886ed94fbfd32417489e909b652a9952/pkgeval/by_hash/3386b42_vs_3559b72/logs/Images/1.6.0-DEV-84648ad804.log